Move Drain-related functions inside the extractor
